alice and bob final

+++ algorithms/graphs/dijkstra.cpp
@@ -3,26 +3,26 @@
 // d gives distances to each point, p gives parent in path
 // from cp-algorithms.com
 
-const long long INF = 10000000000000000LL;
-void dijkstra(int s, vector<long long> & d, vector<int> & p, vector<vector<pair<int,long long>>> &adj) {
+const ll INF = 10000000000000000LL;
+void dijkstra(int s, vector<ll>& d, vector<int>& p, vector<vector<pair<int,ll>>>& adj) {
     int n = adj.size();
     d.assign(n, INF);
     p.assign(n, -1);
 
     d[s] = 0;
-    using pii = pair<long long, int>;
+    using pii = pair<ll, int>;
     priority_queue<pii, vector<pii>, greater<pii>> q;
     q.push({0, s});
     while (!q.empty()) {
         int v = q.top().second;
-        long long d_v = q.top().first;
+        ll d_v = q.top().first;
         q.pop();
         if (d_v != d[v])
             continue;
 
         for (auto edge : adj[v]) {
             int to = edge.first;
-            long long len = edge.second;
+            ll len = edge.second;
 
             if (d[v] + len < d[to]) {
                 d[to] = d[v] + len;
@@ -56,24 +56,24 @@ Time complexity same as dijsktra O(E log V)
 Tested (directed) on cses/additional/visiting_cities
 Tested (undirected) on https://leetcode.com/problems/number-of-ways-to-arrive-at-destination/
 */
-const long long INF = 10000000000000000LL;
+const ll INF = 10000000000000000LL;
 void dijkstra(int s, VLL& d, vector<vector<pair<int,ll>>>& adj) {
     int n = adj.size();
     d.assign(n, INF);
     d[s] = 0;
-    using pii = pair<long long, int>;
+    using pii = pair<ll, int>;
     priority_queue<pii, vector<pii>, greater<pii>> q;
     q.push({0, s});
     while (!q.empty()) {
         int v = q.top().second;
-        long long d_v = q.top().first;
+        ll d_v = q.top().first;
         q.pop();
         if (d_v != d[v])
             continue;
 
         for (auto edge : adj[v]) {
             int to = edge.first;
-            long long len = edge.second;
+            ll len = edge.second;
 
             if (d[v] + len < d[to]) {
                 d[to] = d[v] + len;
